Skip processing if output is not to a terminal
We only actually need the paging if the output is printed on the screen. When smartless is redirected into a file (cmd | smartless > /tmp/file), or used inside a pipe (cmd | smartless | cmd), the processing / duplicate output actually can do harm / confuse the user. Of course, nobody would intentionally put a (useless) pager inside a pipe, but some scripts unconditionally use a pager for all their output, and the user may redirect that script's output. Or the user recalls a previous (paged) command and appends additional processing (!! | head).
If stdout is not to a terminal, simply pass all input to cat, which will simply pass it through, and then end the script.