Print the first lines to stdout, not stderr
The user may do a redirection of stderr (on the whole pipeline including smartless), and now the output does not appear in the terminal. As the pager uses stdout, so should the mirrored output for the terminal. And some terminals are set up in a way to present stderr output in a different color.