ENH: Allow to capture full output that does not fit terminal
When the pager is invoked and it uses the alternate screen, there will be no history of the (browsed) output after paging quits. smartless will only mirror the first lines in the terminal. It can be useful to keep a full record of the paged file, and disk space today is cheap and abundant.
Allow to capture the full contents of paged files in uniquely named files inside the temp directory; the filespec is included in the smartless message about switching to the pager. Use of the temp file enables eventual cleanup. The prefix of the capture files can be configured in SMARTLESS_CAPTURE_PREFIX; this then also turns on the feature.