Use rdfind (if installed) to de-duplicate previous identical captures
If the same command output is paged / the same file viewed, the captured output will be identical. We cannot remove the capture file itself, as its filespec has been written to the terminal as a reference for the user, but we can replace identical files with hard links. This is what the rdfind tool (which is available in most Linux distribution repositories) does.
Launch the tool in the background, pass it all smartless captures in the temp directory.