FIX: Previous solution duplicates last line when it's exactly the last directly displayed one
Instead, we need to omit the trailing newline from the last printed directly displayed line (so that the pager stuff can immediately append its "more shown in" message), and only add the newline if no paging happens.
Furthermore, incomplete lines can both be read from the initial loop, and also from the paging check. Rename the first to earlyLine so that it doesn't get clobbered by the second, and print either at the end if no paging happens.