Replace timeout reads with empty line marker that signals end of current wrapped line
The problem with this coproc is that one input line yields an unknown number (exactly what we're interested here) of wrapped output. So far, we've used a read timeout to (hopefully) get all of the wrapped lines, but this is noticeably slow, and the timing relies on the speed of the interprocess communication.
Switch to a marker that is included in the input sent, and then located in the processed output. Wrapping of a single non-empty line will never result in an empty line on its own, so a trailing empty line is the ideal marker here.