CHG: Don't mention the original pager, use mr alias and capture filename for recall
It's of little interest what program got used to show the full contents (assuming this also happened in the terminal; i.e. didn't fork off a GUI program). So just show "(full output truncated)".
For the capture, the entire path is long, making the whole message long (and often broken across lines, which makes it difficult to copy). We can reuse the mr / rl aliases and the capture filename for a shorter recall.
Prepend a $ to indicate the shell command (and allow me to use my Tmux prefix + Alt-S mapping to capture it), omit the parentheses also for ease of capturing.