CHG: Print the newline with the last read line and put the recall message on a new line
It's been mildly annoying that the recall message is appended to the last shown line, and so the recall command is potentially broken over two lines (which makes it harder to capture or even retype). But this also means that a command that in parallel directly updates the tty (like my invocationTimer for the "todo.sh renderrelativemarkers") also does this at the end of that line, although without paging it would do this at the beginning of the next line.
This also simplifies the implementation by avoiding the special case for $n -eq 0.