FIX: An overly long line at the start / just before the pager is invoked is shown in full
The while loop doesn't account for the (wrapped) multi-line output of the current line, only of previous lines.
Move the countNetLines() into the loop and only accumulate complete shown lines in lines, renamed to shownCompleteLines.
If the current line is too long, pass it through the pager in a new lineRemainder variable.
Same needs to happen for an incomplete line without newline.
Paging has to use either $lineRemainder or read a new input line; include the newline in those variables and remove the explicit newlines inside the loop.