Don't repeat content already shown from the beginning
We're already keeping track of what got printed: shownCompleteLines + (optionally) netLines. Use wc to count the bytes, and read the captured contents from that position onwards.
This completely avoids an overlap in the contents; however, it does not detect if both begin and end of the contents have been fully shown, so the (full) contents are still interrupted by the truncation message in the middle. But that's okay as it's rare and tolerable (we're not lying to the user).