ENH: Kill the pager if the wrapped content can be fully displayed
The pager is now already invoked when half of the available terminal lines have been used - the truncation message now is placed in the middle.
We cannot further delay the pager; (slow) line updates would not be shown or would have to be overwritten again in the terminal.
But if on further reading of the complete input it turns out that paging wasn't necessary because the whole wrapped contents do fit the SMARTLESS_NUM_LINES of the terminal, the pager is useless. (Well, it offers easier naviation with searches.)
We can check for that by injecting another pipeline step right before the pager (so it receives the whole contents, like the pager). That killUselessPager() uses a separate invocation of nowrap (as the coproc only handles individual lines, and reading, feeding, and aggregating those likely is slower than a separate process) to count the effective wrapped lines, and kills the pager (whose process ID is unknown, as it's started by a later pipeline step, but we can use the script's PID to narrow it down).