BUG: Regression: Don't add a newline after netLines
There's no newline in the input, so no newline must happen in the output so that there's no additional empty line when the output is shown in full.
To ensure that the truncation message always appears at the beginning of a fresh line, we need to add a newline if the head content had netLines; i.e. wasn't concluded with a newline itself. As netLines is overwritten by the countNetLines() invocations in handleTail(), the content as it was after captureHead() is stored in alreadyShownNetlines.