BUG: Wrapping of reversed text doesn't consider tab position correctly
Truncation so far reversed both lines and characters, but undid the reversal of the ANSI escape sequences, so that those are properly accounted for. Unfortunately, the tab character is rendered with differing widths based on its position, and the reversal affects the position.
Instead, what we need to do is still reverse both lines and characters (to consume as much as possible from the rear end), but for each processed input line, unreverse it in countNetLines(). In proper text order, this means that tab characters are properly rendered, and it also has the nice side effect that the reversal of the ANSI escape sequences isn't necessary any longer.
As countNetLines() is used both for head and tail, and all of the above only applied to the tail, we however need to distinguish between both modes now. To avoid invoking "rev" for each line separately, this needs to be included in the coproc, anyway; the new nowrapReverser is set up by handleTail() (and the coproc used by handleHead() is closed and re-opened).
Unfortunately, the processing of netLines is also affected by the mode: Due to the reversal, we now need to take the passed count of available lines from the end for handleTail(). The implementation changes from direct concatenation until the count is reached to accumulating in an array and then taking the appropriate slice.