XXX: less eats the head's last char right at the end of the line
,----[ demo ]----
| $ printf 'foobar%74s' XY; less -f /dev/null; printf 'LALA\n'
| foobar                                                                        XL
| ALA
| $ printf 'foobar%74s' XY; : less -f /dev/null; printf 'LALA\n'
| foobar                                                                        XY
| LALA
`----

When the cursor is positioned in the last column and less (551 (GNU regular expressions)) is invoked, the last character gets erased; following output (i.e. after less) overwrites that last character instead of continuing in column 1 of the following line. This is regardless of which physical line; i.e. it is not related to scrolling. If there's still space to the right, the problem doesn't occur.
The problem can be avoided by printing a newline, so that the cursor is already at the beginning of the following line. We'll definitely print the remainder of the input (possibly after a truncation message), anyway.
That additional newline has to be omitted from the truncation message. It's only printed when there are $alreadyShownNetlines; introduce a new variable $headTailSeparator and clear it in that case.
We must only print the newline if we have truncation and when the truncation happens in the middle of an input line; i.e. when $newLineCnt is greater than 1.
When the entire input fits the terminal without a truncation message, the additional newline has no effect; the output would continue at column 1 of the following line in any case.